Creating an Ai movement system.
In this commit, there are two types of Ai systems one that is very simple and the other that is able to do a bit more.

The first Ai system is able to patrol a set of coordinates in a continuous loop.

The second Ai system is able to do one of two things, it is able to either patrol a set of coordinates or follow the player if it sees him when patrolling.

In future Commits/iterations of this project, I will be refining the project and deleting redundant systems. Additionally, I will be improving the Ai system by making a more complex Ai by adding more functionality and in the future, I will possibly be adding animations to the Ai.